+++ cmd/app/http.go
@@ -26,10 +26,10 @@ import (
 
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/sigstore/fulcio/pkg/api"
 	gw "github.com/sigstore/fulcio/pkg/generated/protobuf"
 	legacy_gw "github.com/sigstore/fulcio/pkg/generated/protobuf/legacy"
 	"github.com/sigstore/fulcio/pkg/log"
+	"github.com/sigstore/fulcio/pkg/server"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/metadata"
@@ -43,7 +43,7 @@ type httpServer struct {
 
 func extractOIDCTokenFromAuthHeader(ctx context.Context, req *http.Request) metadata.MD {
 	token := strings.Replace(req.Header.Get("Authorization"), "Bearer ", "", 1)
-	return metadata.Pairs(api.MetadataOIDCTokenKey, token)
+	return metadata.Pairs(server.MetadataOIDCTokenKey, token)
 }
 
 func createHTTPServer(ctx context.Context, serverEndpoint string, grpcServer, legacyGRPCServer *grpcServer) httpServer {
@@ -63,9 +63,9 @@ func createHTTPServer(ctx context.Context, serverEndpoint string, grpcServer, le
 	}
 
 	// Limit request size
-	handler := api.WithMaxBytes(mux, maxMsgSize)
-	handler = promhttp.InstrumentHandlerDuration(api.MetricLatency, handler)
-	handler = promhttp.InstrumentHandlerCounter(api.RequestsCount, handler)
+	handler := server.WithMaxBytes(mux, maxMsgSize)
+	handler = promhttp.InstrumentHandlerDuration(server.MetricLatency, handler)
+	handler = promhttp.InstrumentHandlerCounter(server.RequestsCount, handler)
 
 	api := http.Server{
 		Addr:    serverEndpoint,
@@ -96,20 +96,20 @@ func setResponseCodeModifier(ctx context.Context, w http.ResponseWriter, _ proto
 	}
 
 	// set SCT if present ahead of modifying response code
-	if vals := md.HeaderMD.Get(api.SCTMetadataKey); len(vals) > 0 {
-		delete(md.HeaderMD, api.SCTMetadataKey)
+	if vals := md.HeaderMD.Get(server.SCTMetadataKey); len(vals) > 0 {
+		delete(md.HeaderMD, server.SCTMetadataKey)
 		delete(w.Header(), "Grpc-Metadata-sct")
 		w.Header().Set("SCT", vals[0])
 	}
 
 	// set http status code
-	if vals := md.HeaderMD.Get(api.HTTPResponseCodeMetadataKey); len(vals) > 0 {
+	if vals := md.HeaderMD.Get(server.HTTPResponseCodeMetadataKey); len(vals) > 0 {
 		code, err := strconv.Atoi(vals[0])
 		if err != nil {
 			return err
 		}
 		// delete the headers to not expose any grpc-metadata in http response
-		delete(md.HeaderMD, api.HTTPResponseCodeMetadataKey)
+		delete(md.HeaderMD, server.HTTPResponseCodeMetadataKey)
 		delete(w.Header(), "Grpc-Metadata-X-Http-Code")
 		w.WriteHeader(code)
 	}
