+++ pkg/server/legacy_server.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 
-package api
+package server
 
 import (
 	"context"
