+++ pkg/server/max_bytes_test.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 
-package api
+package server
 
 import (
 	"io/ioutil"
